// packages/server/src/types/user.ts
interface Adress {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

interface Geo {
  lat: string;
  lng: string;
}

interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  password: string;
  permission: string;
  adress: Adress;
  phone: string;
  website: string;
  company: Company;
}